requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
langgraph
pyppeteer
langchain-tavily
elasticsearch
mem0ai
mcp